requirements.txt
# --extra-index-url https://download.pytorch.org/whl/test/cu124

numpy
matplotlib
pydicom
pylidc
scipy
pyyaml
tqdm
setuptools
torch
torchvision
albumentations